test: move dhis-service-validation tests into dhis-test-integration
